xorg: Libinput -> Libinut in Xorg
libinput is a library to handle input devices in Wayland compositors and
to provide a generic X.Org input driver. It provides device detection,
device handling, input device event processing and abstraction so
minimize the amount of custom input code compositors need to provide the
common set of functionality that users expect.
https://wiki.archlinux.org/index.php/libinput